Actions, Roles and Reference cards
read the txt file to get a sense of what is there, I tryed to compare
these to other parts of code to make it clearer. basically this is a
structure that fits the rest but I have not modified any code to include
them. i did not want to mess arround and break the rest.